### 2025-07-28 #8658 @editorialbot
[PRE REVIEW]: Gw-detect-power: A Python Package to predict the detectability of land management changes on freshwater contaminants despite lag.
**Submitting author:** @hansonmcoombs (<a href="http://orcid.org/0009-0007-2974-0335">Matt  Dumont</a>)
**Repository:** https://github.com/Komanawa/komanawa-gw-detect-power
**Branch with paper.md** (empty if default branch): development
**Version:** 2.0.2
