### 2025-07-29 #8658 @kthyng
@hansonmcoombs perhaps something is going on with your paper. Not sure why the automatic processing keeps stopping. I don't have time to check on it now but if you have a chance to dig in, here is some info:

Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it. Then use the command `@editorialbot generate pdf` after making changes to the .md file to make a new PDF. editorialbot commands need to be the first entry in a new comment.
