### 2025-10-16 #8658 @hansonmcoombs
@kanishkan91

> Going through your repo it seems this package has commits froma single person. Can you explain the role of the other co-authors in development?

The other author Mike Kitteridge provided base code prior to moving this package onto github.  He also provided design guidance, editorial, and code review support.

> Also, your repo does not have contribution instructions.

I have fixed this.

> Can you explain whether this software has been used in any ongoing or published research projects?

It has been used to conduct a New Zealand wide analysis of the detection power of current monitoring programmes for groundwater (https://doi.org/10.1016/j.scitotenv.2024.171759) and to undertake multiple detailed local case studies of specific catchments.

Does this address these issues satisfactorally?
