### 2025-10-16 #8658 @kanishkan91
@jsta would you be interested in reviewing this paper and software package for JOSS?
