### 2025-12-14 #9390 @hansonmcoombs
@hassaniazi  Thanks for the review. A quick question for your requested quickstart guide (which I'm very happy to provide):

Before I run off in the wrong direction.  For a quickstart guide are you after something like a Juyperter notebook with a full worked example and if so how does this differ from the examples in: https://github.com/Komanawa/komanawa-gw-detect-power-worked-examples/blob/main/Whakauru_Stream_worked_example.ipynb  Also note that I have added a pure python version of the worked examples.

Noteing that I agree with you re pulling them into the main repo.
