### 2025-12-15 614f5f7
bring worked examples into the repo (instead of separate repo) add a bit more of a landing page for the repo) including space for requested quickstart.
