+++ worked_examples/Python_only_examples/increasing_run_efficency..py
@@ -1,0 +1,149 @@
+"""
+created matt_dumont 
+on: 16/02/24
+"""
+#%% md
+# Increasing Runtime Efficiency
+
+# The detection power calculations can be slow, especially if you are running a large number of detection power simulations and using a computationally expensive detection method.  This notebook provides some tips on how to increase the runtime efficiency of the detection power calculations.
+
+# Before we get too far let's import the required packages and generate some data to work with.
+#%%
+import numpy as np
+from komanawa.gw_detect_power import AutoDetectionPowerSlope, AutoDetectionPowerCounterFactual
+#%% md
+## Testing run time (Slope)
+
+# The classes have a method to test the run time of the detection power calculations. This can be useful to determine how long it will take to run a large number of simulations.  An example follows:
+#%%
+dpc = AutoDetectionPowerSlope(
+    significance_mode='n-section-mann-kendall', nsims=1000,
+    expect_slope=[1, 0, -1], nparts=3, min_part_size=10, no_trend_alpha=0.50,
+    return_true_conc=True, return_noisy_conc_itters=3)
+
+dpc.time_test_power_calc_itter(
+    testnitter=5,  # only run 5 change detection iterations instead of 1000 as per dpc.nsims
+    # all the following are kwargs for the DetectionPowerCalculator.power_calc function
+    idv='true',
+    error=0.5,
+    mrt_model='binary_exponential_piston_flow',
+    samp_years=10,
+    samp_per_year=10,
+    implementation_time=5,
+    initial_conc=10,
+    target_conc=5,
+    prev_slope=1,
+    max_conc_lim=25,
+    min_conc_lim=1,
+    mrt=5,
+    #
+    mrt_p1=3,
+    frac_p1=0.7,
+    f_p1=0.7,
+    f_p2=0.7,
+    #
+    seed=558
+)
+#%% md
+## Testing run time (Counterfactual)
+
+# The same method is available for the counterfactual detection power calculations.  An example follows:
+#%%
+dpc_counter = AutoDetectionPowerCounterFactual(significance_mode='paired-t-test', nsims=1000,
+                                           p_value=0.05,
+                                           min_samples=10,
+                                           alternative='alt!=base')
+
+dpc_counter.time_test_power_calc_itter(
+    testnitter=5,  # only run 5 change detection iterations instead of 1000 as per dpc.nsims
+    # all the following are kwargs for the DetectionPowerCalculator.power_calc function
+    idv='true',
+    error_base=0.5,
+    mrt_model='binary_exponential_piston_flow',
+    samp_years=10,
+    samp_per_year=10,
+    implementation_time_alt=5,
+    initial_conc=10,
+    target_conc_alt=5,
+    prev_slope=1,
+    max_conc_lim=25,
+    min_conc_lim=1,
+    mrt=5,
+    #
+    mrt_p1=3,
+    frac_p1=0.7,
+    f_p1=0.7,
+    f_p2=0.7,
+    #
+    seed=558
+)
+#%% md
+## Running in muliprocessing mode
+
+# The classes are set up to quickly run multiple detection power estimates in parallel. Essentially you are expected to pre-process the inputs and then pass the runs. Most kwargs can be passed as either an array or as a single value  An example follows:
+#%%
+dpc = AutoDetectionPowerSlope(
+    significance_mode='linear-regression', nsims=1000,
+    ncores=3  # set the number of cores to use
+)
+
+outdata = dpc.mulitprocess_power_calcs(outpath=None,  # can be saved to a .hdf file if desired
+                                       idv_vals=np.array([f'run_{i}' for i in range(6)]),
+                                       error_vals=0.5,
+                                       samp_years_vals=np.array([5, 5, 5, 10, 10, 10]),
[…]
