### 2025-12-15 09cebaf
update type hinting
