+++ src/komanawa/gw_detect_power/change_detection_counterfactual.py
@@ -3,11 +3,15 @@ created matt_dumont
 on: 25/01/24
 """
 from pathlib import Path
+
+import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 from scipy import stats
 import warnings
 import logging
+from numbers import Number
+from typing import Union, Optional, Tuple, Literal, Sequence
 from komanawa.gw_detect_power.base_detection_calculator import BaseDetectionCalculator
 
 
@@ -62,11 +66,11 @@ class DetectionPowerCounterFactual(BaseDetectionCalculator):
     _scipy_alternatives = ('two-sided', 'less', 'greater')
 
     def __init__(self,
-                 significance_mode,
-                 nsims=1000,
-                 p_value=0.05,
-                 min_samples=10,
-                 alternative='alt!=base',
+                 significance_mode: Literal['paired-t-test', 'wilcoxon-signed-rank-test'],
+                 nsims: int = 1000,
+                 p_value: float = 0.05,
+                 min_samples: int = 10,
+                 alternative: Literal['alt!=base', 'alt<base', 'alt>base'] = 'alt!=base',
                  wx_zero_method='wilcox', wx_correction=False, wx_method='auto',
                  ncores=None,
                  log_level=logging.INFO,
@@ -122,7 +126,8 @@ class DetectionPowerCounterFactual(BaseDetectionCalculator):
         self.log_level = log_level
         self.significance_mode = significance_mode
 
-    def plot_iteration(self, y0_base, y0_alt, true_conc_base, true_conc_alt, ax=None):
+    def plot_iteration(self, y0_base: Sequence, y0_alt: Sequence, true_conc_base: Sequence, true_conc_alt: Sequence,
+                       ax: Optional[plt.Axes] = None) -> Tuple[plt.Figure, plt.Axes]:
         """
         plot the concentration data itteration and the true concentration data
 
@@ -257,12 +262,12 @@ class DetectionPowerCounterFactual(BaseDetectionCalculator):
     def power_calc(self, idv, error_base: float,
                    true_conc_base: np.ndarray,
                    true_conc_alt: np.ndarray,
-                   error_alt: {float, None} = None,
-                   seed_base: {int, None} = None,
-                   seed_alt: {int, None} = None,
+                   error_alt: Optional[float] = None,
+                   seed_base: Optional[int] = None,
+                   seed_alt: Optional[int] = None,
                    testnitter=None,
                    **kwargs
-                   ):
+                   )-> Union[pd.Series, dict]:
 
         """
         calculate the counterfactual detection power of a pair of concentration time series note the power is calculated using the sampling frequency of the true_conc_base/alt, if you want to test the power at a different sampling frequency then you should resample the true_conc_base/alt before passing it to this function
@@ -301,15 +306,15 @@ class DetectionPowerCounterFactual(BaseDetectionCalculator):
             self,
             outpath: {Path, None, str},
             idv_vals: np.ndarray,
-            true_conc_base_vals: {np.ndarray, list},
-            true_conc_alt_vals: {np.ndarray, list},
-            error_base_vals: {np.ndarray, None, float},
-            error_alt_vals: {np.ndarray, None, float} = None,
-            seed_alt_vals_vals: {np.ndarray, int, None} = None,
-            seed_base_vals_vals: {np.ndarray, int, None} = None,
+            true_conc_base_vals: Sequence,
+            true_conc_alt_vals: Sequence,
+            error_base_vals: Union[Sequence, float],
+            error_alt_vals: Optional[Union[Sequence, float]] = None,
+            seed_alt_vals_vals: Optional[Union[Sequence[int], int]] = None,
[…]

[…]
