### 2025-12-15 #9390 @hansonmcoombs
Re:  "CODE: Some warnings appear when I run test_slope_detector.py"

These warnings are:

- DeprecationWarning: This process (pid=3432892) is multi-threaded, use of fork() may lead to deadlocks in the child. --> fine for now
- RuntimeWarning: Mean of empty slice --> expected behaviour, I have now suppressed this warning for the tests.
- MatplotlibDeprecationWarning: The get_cmap function was deprecated in Matplotlib 3.7 --> already handled in upstream package via try/except loop so it'll work once depreciated.
- The behavior of DataFrame concatenation with empty or all-NA entries is deprecated. -->  fixed in development.

I also took the oppertunity to update the type hinting.

I'll pull to main (and update PyPI) when the review is done.
