### 2025-12-15 dbd020b
start working through some JOSS review comments
