+++ review_todos.txt
@@ -1,0 +1,62 @@
+
+# todo delete when done with review
+This holds the comments made by hassaniazi as a working document for things to address...
+
+
+Overview comment
+===============================================
+
+The paper/package would benefit from the inclusion of a quick-starter notebook to demonstrate a complete example run case and add corresponding text in the manuscript. This should include data loading, parameter specification, model execution sequencing, and example outputs and interpretation of outputs. Providing such a notebook would substantially improve accessibility, reproducibility, and usability for new users.  # todo --> asked for clarification on what is wanted here
+
+
+
+Manuscript -- # todo
+===============================================
+Line 11 - Correct in-text citation format
+Line 26 - Writing could be improved, e.g., the opening sentence could be restructured for broader appeal and ease to follow: “Reducing the loss of nutrients (e.g., nitrogen) from land into surface and groundwater environments takes considerable time and financial investment...”. Starting from "land-based mitigation" might tune out people not in your area of expertise. Similar stylistic refinements would improve clarity throughout the manuscript.
+Line 34 - The manuscript should improve the flow of the argument to more clearly articulate the need, e.g., starting from why detecting changes is important. Making the connections clear and making them flow logically is important, e.g., nutrient losses from land to freshwater contamination, changes in freshwater contaminant levels to water quality and ecosystem health, the challenges of detecting these changes due to time lags, monitoring costs, and uncertainty, resulting in the need for rapid, robust, and standardized tools, hence the package… Making this logical progression explicit would better justify the development and application of the package.
+In its current form, the paper does not provide sufficient detail to function as a comprehensive software or methods publication. The authors should substantially expand the description of: - broader theory, background, and practices currently in the area; the underlying methods and statistical assumptions
+overall code architecture and design rationale; key modules and files within the package
+typical workflows and use cases; expected inputs and outputs (including data formats)
+installation procedures and dependencies, and usability considerations for different user groups
+visual outputs or screenshots illustrating typical results
+These additions would significantly strengthen the paper’s value as a reference for applied researchers and practitioners.
+
+Code -- DONE
+===============================================
+
+Below are the comments on unchecked items:
+===============================================
+
+Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
+------------------------------------------------
+
+    I did not find any data but think it’s important to provide some for example cases. I do see test data.
+
+Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
+------------------------------------------------
+
+    Hard to verify reproducibility without data or example cases. Tests do run.
+
+Functionality: Have the functional claims of the software been confirmed?
+------------------------------------------------
+
+    Same point as reproducibility
+
+A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
+------------------------------------------------
+
+    Statement of need not present in documentation.
+
+Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
+------------------------------------------------
+
+    Absent. Suggest including
+
+Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
+------------------------------------------------
+    Missing. I am interpreting it as the sequence of launching scripts. Suggest adding it to the documentation as well, along with the example.  # todo asked for details from reviewer
+
+State of the field: Do the authors describe how this software compares to other commonly-used packages?
+------------------------------------------------
+    Not present in the paper.
